#2 feat: SchoolTypeDto 선언
