+++ src/main/java/com/autoever/recall/school/domain/SchoolType.java
@@ -1,34 +1,7 @@
 package com.autoever.recall.school.domain;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonValue;
-
-import java.util.Arrays;
-
 public enum SchoolType {
-    ELEMENTARY("elementary"),
-    MIDDLE("middle"),
-    HIGH("high");
-
-    private final String key;
-
-    SchoolType(String key) {
-        this.key = key;
-    }
-
-    // Response: 객체를 JSON으로 보낼 때 'key'값을 사용함
-    @JsonValue
-    public String getKey() {
-        return key;
-    }
-
-    // Request: JSON 문자열을 Enum으로 바꿀 때 이 메서드를 사용함
-    @JsonCreator
-    public static SchoolType fromKey(String key) {
-        return Arrays.stream(SchoolType.values())
-                     .filter(role -> role.getKey().equals(key))
-                     .findFirst()
-                     .orElseThrow(() -> new IllegalArgumentException("유효하지 않은 학교 유형(SchoolType)입니다: " + key));
-    }
-
+    ELEMENTARY,
+    MIDDLE,
+    HIGH;
 }
