+++ src/main/java/com/autoever/recall/school/dto/SchoolDto.java
@@ -1,13 +1,12 @@
 package com.autoever.recall.school.dto;
 
 import com.autoever.recall.school.domain.School;
-import com.autoever.recall.school.domain.SchoolType;
 
 import java.time.LocalDateTime;
 
 public record SchoolDto(
         Long id,
-        SchoolType type,
+        String type,
         String name,
         String imageUrl,
         String address,
@@ -16,7 +15,7 @@ public record SchoolDto(
     public static SchoolDto from(School school) {
         return new SchoolDto(
                 school.getId(),
-                school.getType(),
+                SchoolTypeDto.from(school.getType()),
                 school.getName(),
                 school.getImageUrl(),
                 school.getAddress(),
